requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-celery-beat==2.4.0
django-crispy-forms==1.14.0
django-cuser==2017.3.16
django-modeltranslation==0.18.4
django-rosetta==0.9.8
django-timezone-field==5.0
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
idna==3.4
iniconfig==1.1.1
kombu==5.2.4
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
polib==1.1.1
prompt-toolkit==3.0.32
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
